+++ package/thunderstore/ExtraSlots/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.0.18
+* new config option for stack size Color of slots in equipment panel
+* quick slots hotbars will no longer overlap map window
+
 # 1.0.17
 * new config option to use several hotbar items at once
 * changed the logic of checking quick slots and utility slots activity. If only one of the values ​​(global key or discovered item) is specified - checking for the unfilled one will not be performed